fix avent fav et inscrit, debut c_create_avent

+++ code/DATABASE/DB_AVent_creation.sql
@@ -12,7 +12,7 @@ CREATE TABLE brand
 (
     brand_id INT PRIMARY KEY AUTO_INCREMENT,
     brand_name VARCHAR(255),
-    siret INT
+    siret BIGINT
 );
 
 CREATE TABLE genre
